feat: Split ObjectListing component into smaller components
