+++ src/components/objectListing/objectListing.component.tsx
@@ -3,10 +3,21 @@ import {
   getCoreRowModel,
   useReactTable,
   VisibilityState,
+  Table as ReactTable,
 } from "@tanstack/react-table";
 import clsx from "clsx";
-import { useEffect, useState, useMemo, useRef, useCallback } from "react";
+import {
+  useEffect,
+  useState,
+  useMemo,
+  useRef,
+  useCallback,
+  forwardRef,
+  Ref,
+  RefObject,
+} from "react";
 import { useVirtual } from "react-virtual";
+import { useDebounce } from "use-debounce";
 
 import { Spinner } from "src/components/icons";
 import { OBJECT_LIST_TABLE } from "src/constants/skylark";
@@ -41,65 +52,36 @@ export interface ObjectListProps {
   isDragging?: boolean;
 }
 
-export const ObjectList = ({
-  withCreateButtons,
-  withObjectSelect,
-  withObjectEdit = false,
+const ObjectSearchResults = ({
+  sortedHeaders,
+  columnVisibility,
+  isPanelOpen,
   panelObject,
   setPanelObject,
   isDragging,
-  isPanelOpen,
-}: ObjectListProps) => {
-  const [searchQuery, setSearchQuery] = useState("");
-  const { objectTypes } = useSkylarkObjectTypes(true);
-  const [searchFilters, setSearchFilters] = useState<SearchFilters>({
-    objectTypes: null,
-    language: null,
-  });
-
+  searchData,
+  withObjectSelect,
+  withObjectEdit,
+  fetchNextPage,
+}: {
+  fetchNextPage?: () => void;
+  searchData: ParsedSkylarkObject[];
+  sortedHeaders: string[];
+  columnVisibility: VisibilityState;
+} & ObjectListProps) => {
   const tableContainerRef = useRef<HTMLDivElement>(null);
-
-  const {
-    data: searchData,
-    error: searchError,
-    isLoading: searchLoading,
-    totalHits,
-    properties,
-    query: graphqlSearchQuery,
-    variables: graphqlSearchQueryVariables,
-    hasNextPage,
-    isFetchingNextPage,
-    isRefetching: searchRefetching,
-    refetch,
-    fetchNextPage,
-  } = useSearch(searchQuery, searchFilters);
-
-  // Sorts objects using the preference array above, any others are added to the end randomly
-  const sortedHeaders = useMemo(() => {
-    const orderedKeysThatExist =
-      properties?.filter((property) => orderedKeys.includes(property)) || [];
-
-    const orderedProperties =
-      properties?.filter((property) => !orderedKeys.includes(property)) || [];
-
-    return [...hardcodedColumns, ...orderedKeysThatExist, ...orderedProperties];
-  }, [properties]);
+  const [rowInEditMode, setRowInEditMode] = useState("");
 
   const parsedColumns = useMemo(
     () =>
       createObjectListingColumns(sortedHeaders, hardcodedColumns, {
         withObjectSelect,
-      }),
+      }) as ColumnDef<object, ParsedSkylarkObject>[],
     [sortedHeaders, withObjectSelect],
   );
 
-  const [rowInEditMode, setRowInEditMode] = useState("");
-  const [columnVisibility, setColumnVisibility] = useState<VisibilityState>(
-    Object.fromEntries(sortedHeaders.map((header) => [header, true])),
-  );
-
   const formattedSearchData = useMemo(() => {
-    const searchDataWithDisplayField = searchData?.map((obj) => {
+    const searchDataWithDisplayField = searchData.map((obj) => {
       return {
         ...obj,
         // When the object type is an image, we want to display its preview in the images tab
@@ -133,23 +115,15 @@ export const ObjectList = ({
         const { scrollHeight, scrollTop, clientHeight } = containerRefElement;
         // once the user has scrolled within 500px of the bottom of the table, fetch more data if there is any
         if (
+          fetchNextPage &&
           searchData.length > 0 &&
-          scrollHeight - scrollTop - clientHeight < 500 &&
-          hasNextPage &&
-          !searchLoading &&
-          !isFetchingNextPage
+          scrollHeight - scrollTop - clientHeight < 500
         ) {
           fetchNextPage();
         }
       }
     },
-    [
-      searchData.length,
-      hasNextPage,
-      searchLoading,
-      isFetchingNextPage,
-      fetchNextPage,
-    ],
+    [searchData.length, fetchNextPage],
   );
 
   // a check on mount and after a fetch to see if the table is already scrolled to the bottom and immediately needs to fetch more data
@@ -159,10 +133,8 @@ export const ObjectList = ({
 
   const table = useReactTable({
     debugAll: false,
-    data: formattedSearchData || [],
-    columns: searchData
-      ? (parsedColumns as ColumnDef<object, ParsedSkylarkObject>[])
-      : [],
+    data: formattedSearchData,
+    columns: parsedColumns,
     getCoreRowModel: getCoreRowModel(),
     columnResizeMode: "onChange",
     state: {
@@ -170,7 +142,7 @@ export const ObjectList = ({
     },
     meta: {
       rowInEditMode,
-      withObjectEdit,
+      withObjectEdit: !!withObjectEdit,
       onEditClick(rowId) {
         setRowInEditMode(rowId);
       },
@@ -179,6 +151,74 @@ export const ObjectList = ({
       },
     },
   });
+
+  return (
+    <div
+      className={clsx(
+        isDragging ? "overflow-hidden" : "overflow-x-auto",
+        "relative mb-6 flex w-full flex-auto flex-grow flex-col overscroll-none md:-ml-4",
+      )}
+      ref={tableContainerRef}
+      data-testid="table-container"
+      onScroll={(e) => fetchMoreOnBottomReached(e.target as HTMLDivElement)}
+    >
+      <Table
+        ref={tableContainerRef}
+        table={table}
+        withCheckbox={withObjectSelect}
+        isLoadingMore={!!fetchNextPage}
+        activeObject={panelObject || undefined}
+        setPanelObject={setPanelObject}
+        withDraggableRow={!!isPanelOpen}
+      />
+    </div>
+  );
+};
+
+export const ObjectSearch = (props: ObjectListProps) => {
+  const { withCreateButtons, setPanelObject, isPanelOpen } = props;
+
+  const [searchQuery, setSearchQuery] = useState("");
+  const { objectTypes } = useSkylarkObjectTypes(true);
+  const [searchFilters, setSearchFilters] = useState<SearchFilters>({
+    objectTypes: null,
+    language: null,
+  });
+  const [debouncedSearchFilters] = useDebounce(searchFilters, 300);
+
+  const {
+    data: searchData,
+    error: searchError,
+    isLoading,
+    totalHits,
+    properties,
+    query: graphqlSearchQuery,
+    variables: graphqlSearchQueryVariables,
+    hasNextPage,
+    isFetchingNextPage,
+    isRefetching: searchRefetching,
+    refetch,
+    fetchNextPage,
+  } = useSearch(searchQuery, debouncedSearchFilters);
+
+  const isSearching =
+    isLoading ||
+    JSON.stringify(searchFilters) !== JSON.stringify(debouncedSearchFilters);
+
+  // Sorts objects using the preference array above, any others are added to the end randomly
+  const sortedHeaders = useMemo(() => {
+    const orderedKeysThatExist =
+      properties?.filter((property) => orderedKeys.includes(property)) || [];
+
+    const orderedProperties =
+      properties?.filter((property) => !orderedKeys.includes(property)) || [];
+
+    return [...hardcodedColumns, ...orderedKeysThatExist, ...orderedProperties];
+  }, [properties]);
+
+  const [columnVisibility, setColumnVisibility] = useState<VisibilityState>(
+    Object.fromEntries(sortedHeaders.map((header) => [header, true])),
+  );
 
   useEffect(() => {
     if (
@@ -218,15 +258,6 @@ export const ObjectList = ({
 
   if (searchError) console.error("Search Errors:", { searchError });
 
-  const { rows } = table.getRowModel();
-  const rowVirtualizer = useVirtual({
-    parentRef: tableContainerRef,
-    size: rows.length,
-    estimateSize: useCallback(() => 40, []),
-    overscan: 35,
-  });
-  const { virtualItems: virtualRows, totalSize } = rowVirtualizer;
-
   return (
     <div
       className={clsx(
@@ -252,7 +283,7 @@ export const ObjectList = ({
               query: graphqlSearchQuery,
               variables: graphqlSearchQueryVariables,
             }}
-            isSearching={searchLoading || searchRefetching}
+            isSearching={isSearching || searchRefetching}
             onRefresh={refetch}
             onQueryChange={setSearchQuery}
             activeFilters={searchFilters}
@@ -262,7 +293,7 @@ export const ObjectList = ({
           />
           <div className="mt-2 flex w-full justify-start pl-3 md:pl-7">
             <p className="text-xs font-medium text-manatee-400">
-              {searchLoading ? "Loading..." : `${totalHits || 0} results`}
+              {isSearching ? "Loading..." : `${totalHits || 0} results`}
             </p>
           </div>
         </div>
@@ -278,41 +309,32 @@ export const ObjectList = ({
           />
         )}
       </div>
-      <div
-        className={clsx(
-          isDragging ? "overflow-hidden" : "overflow-x-auto",
-          "relative mb-6 flex w-full flex-auto flex-grow flex-col overscroll-none md:-ml-4",
-        )}
-        ref={tableContainerRef}
-        data-testid="table-container"
-        onScroll={(e) => fetchMoreOnBottomReached(e.target as HTMLDivElement)}
-      >
-        {!searchLoading && searchData && (
-          <Table
-            table={table}
-            virtualRows={virtualRows}
-            totalRows={totalSize}
-            withCheckbox={withObjectSelect}
-            isLoadingMore={hasNextPage || isFetchingNextPage}
-            activeObject={panelObject || undefined}
-            setPanelObject={setPanelObject}
-            withDraggableRow={!!isPanelOpen}
-          />
-        )}
-        {(searchLoading || searchData) && (
-          <div className="items-top justify-left flex h-96 w-full flex-col space-y-2 text-sm text-manatee-600 md:text-base">
-            {searchLoading && (
-              <div className="flex w-full justify-center">
-                <Spinner className="h-10 w-10 animate-spin" />
-              </div>
-            )}
+      {!isSearching && (
+        <ObjectSearchResults
+          {...props}
+          fetchNextPage={
+            hasNextPage && !isFetchingNextPage
+              ? () => fetchNextPage()
+              : undefined
+          }
+          sortedHeaders={sortedHeaders}
+          searchData={searchData}
+          columnVisibility={columnVisibility}
+        />
+      )}
+      {(isSearching || (searchData && searchData.length === 0)) && (
+        <div className="items-top justify-left flex h-96 w-full flex-col space-y-2 text-sm text-manatee-600 md:text-base">
+          {isSearching && (
+            <div className="flex w-full justify-center">
+              <Spinner className="h-10 w-10 animate-spin" />
+            </div>
+          )}
 
-            {!searchLoading && searchData && searchData.length === 0 && (
-              <p className="md:ml-6">{`No results containing all your search terms were found.`}</p>
-            )}
-          </div>
-        )}
-      </div>
+          {!isSearching && searchData && searchData.length === 0 && (
+            <p className="md:ml-6">{`No results containing all your search terms were found.`}</p>
+          )}
+        </div>
+      )}
     </div>
   );
 };
